refactor: move TaskItem and start of asana functions
